% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\alias{play_yatzy}
\alias{play_yahtzee}
\alias{play_maxi}
\alias{play_forced}
\alias{play_mitzy}
\title{Play dice game variant}
\usage{
play_yatzy()

play_yahtzee()

play_maxi()

play_forced()

play_mitzy()
}
\description{
Shortcut function calls for calling the variants.
}
\examples{
  play_yatzy()

}
