% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahtzR.R
\name{yahtzR}
\alias{yahtzR}
\title{yahtzR}
\usage{
yahtzR(ruleset = "standard", work.mode = F)
}
\arguments{
\item{ruleset}{Character value. Name of ruleset to apply. Valid options are:

\itemize{
   \item "standard" - Regular "Forced joker" rules.
   \item "4 rolls" - standard rules, but with 4 rolls allowed per turn.
   \item "5 rolls" - standard rules, but with 5 rolls allowed per turn.
}}

\item{work.mode}{Logical. If TRUE, then display the game in a manner which looks a bit
  more like you're actually working!

  N.B. I accept no responsibility if your boss still notices!}
}
\description{
Play the game of yahtzee in an R session
}
\details{
Yahtzee is a classic dice rolling game and, heck, why not play it in R instead of doing some
important machine learning or whatever.

Work mode prints minimal help messages and prompting, but accepts exactly the same inputs
as regular mode. It assumes that you have already played in regular (verbose) mode and
therefore understand what inputs are available at the various prompts.

Type "help" or "?" at any prompt for more information.
}
\examples{
  yahtzR(work.mode=T)

}
