% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahtzR.R
\name{yahtzR}
\alias{yahtzR}
\title{yahtzR}
\usage{
yahtzR(variation = "yahtzee", rolls_per_turn = 3, work.mode = FALSE)
}
\arguments{
\item{variation}{Character value. Name of game variation and ruleset to apply. Valid options are:

\itemize{
   \item "yahtzee" - Official Yahtzee game with "Forced Joker" rules.
         \href{https://en.wikipedia.org/wiki/Yahtzee}{Wikipedia}
   \item "yatzy" - The public domain variation of the game.
         \href{https://en.wikipedia.org/wiki/Yatzy}{Wikipedia}
   \item "forced_yatzy" - Regular Yatzy, but you have to fill the card from top to
         bottom (see previous wiki).
   \item "maxi_yatzy" - Version of Yatzy with 6 dice, and some extra game rules.
         \href{https://en.wikipedia.org/wiki/Yatzy#Maxi_Yatzy}{Wikipedia}
   \item "mitzy" - a simplified, 4 dice game with rules similar to Yatzy.
}}

\item{rolls_per_turn}{Integer value. The number of dice rolls allowed per turn.
Usually (and by default) 3, but can be changed to any other value to create your own bespoke
game variation.}

\item{work.mode}{Logical. If TRUE, then display the game in a manner which looks a bit
  more like you're actually working!

  N.B. I accept no responsibility if your boss still notices!}
}
\description{
Play variations of the well-known dice rolling game in an R session.
}
\details{
Yatzy (and variants) are a classic dice rolling game and, heck, why not play then in
an R console instead of doing some important machine learning or whatever.

Work mode prints minimal help messages and prompting, but accepts exactly the same inputs
as regular mode. It assumes that you have already played in regular (verbose) mode and
therefore understand what inputs are available at the various prompts.

Type "help" or "?" at any prompt for more information.
}
\examples{
  yahtzR(work.mode=T)

  #This is just a silly idea :)
  yahtzR(variation="yatzy", rolls_per_turn = 10)

}
