% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahtzR.R
\name{yahtzR}
\alias{yahtzR}
\title{yahtzR}
\usage{
yahtzR(work.mode = F)
}
\arguments{
\item{work.mode}{Logical. If TRUE, then display the game in a manner which looks like you're working!

  N.B. I accept no responsibility if your boss still notices!}
}
\description{
Play the game of yahtzee in an R session
}
\examples{
  yahtzR(work.mode=T)

}
